[Silver V] Title: 수들의 합 5, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/2018. 수들의 합 5/수들의 합 5.cc
@@ -8,8 +8,8 @@ int main(void){
     cin.tie(NULL);
     cout.tie(NULL);
 
-    long long int N;
-    long long int sum = 0;
+    int N;
+    int sum = 0;
     cin >> N;
     
     for(int i=1;;i+=2){
